ai_experiments/New GEN RecSys 2/JibJobRecSys/requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
torch==2.0.1
torch-geometric==2.3.1
transformers==4.36.2
tqdm==4.65.0
PyYAML==6.0
